+++ packages/editor/package.json
@@ -40,10 +40,7 @@
   },
   "peerDependencies": {
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
-  },
-  "dependencies": {
-    "dompurify": "^3.0.6",
-    "react-sane-contenteditable": "^1.6.1"
+    "react-dom": "^18.2.0",
+    "dompurify": "^3.0.6"
   }
 }
