+++ packages/editor/src/index.ts
@@ -1,5 +1,9 @@
 import Editor from "./components/Editor";
-
 import createStyledText from "./components/extendable/createStyledText";
+import factories from "./lib/factories";
 
-export { Editor, createStyledText };
+export {
+  Editor,
+  createStyledText,
+  factories
+};
